alteraçoes

+++ src/main/java/Grupo3pt/iade/ChavesApp/controllers/GameController.java
@@ -21,7 +21,7 @@ public class GameController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<Game> getGameById(@PathVariable Integer id) {
+    public ResponseEntity<Game> getGameById(@PathVariable Long id) {
         return gameService.getGameById(id)
                 .map(ResponseEntity::ok)
                 .orElse(ResponseEntity.notFound().build());
@@ -33,7 +33,7 @@ public class GameController {
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<Game> updateGame(@PathVariable Integer id, @RequestBody Game gameDetails) {
+    public ResponseEntity<Game> updateGame(@PathVariable Long id, @RequestBody Game gameDetails) {
         return gameService.getGameById(id)
                 .map(game -> {
                     Game updatedGame = gameService.updateGame(id, gameDetails);
@@ -43,7 +43,7 @@ public class GameController {
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteGame(@PathVariable Integer id) {
+    public ResponseEntity<Void> deleteGame(@PathVariable Long id) {
         if (gameService.getGameById(id).isPresent()) {
             gameService.deleteGame(id);
             return ResponseEntity.noContent().build();
@@ -51,6 +51,7 @@ public class GameController {
         return ResponseEntity.notFound().build();
     }
 }
+
